fix: route bench execution through recipes
