+++ examples/recipes/bench-plugin.json
@@ -1,0 +1,34 @@
+{
+  "schema": "wp-codebox/workspace-recipe/v1",
+  "runtime": {
+    "wp": "7.0",
+    "blueprint": {
+      "steps": []
+    }
+  },
+  "inputs": {
+    "extraPlugins": [
+      {
+        "source": "../bench-plugin",
+        "slug": "bench-plugin",
+        "activate": false
+      }
+    ]
+  },
+  "workflow": {
+    "steps": [
+      {
+        "command": "wordpress.bench",
+        "args": [
+          "component-id=bench-plugin",
+          "plugin-slug=bench-plugin",
+          "iterations=3",
+          "warmup=1",
+          "dependency-slugs=",
+          "env-json={}",
+          "workloads-json=[]"
+        ]
+      }
+    ]
+  }
+}
